azure-function-sync-orchestrator/requirements.txt
azure-functions
supabase
azure-identity
azure-mgmt-appcontainers
